Move incomplete invoices to invoice model
