+++ app/models/invoice.rb
@@ -8,4 +8,8 @@ class Invoice < ApplicationRecord
   validates_presence_of :customer_id
   validates_presence_of :status
   validates_numericality_of :customer_id
+
+  def self.incomplete_invoices
+    select(:invoice_id).distinct.where.not(status: :shipped).pluck(:invoice_id)
+  end
 end
